[3787] Cleanup extension functions
