+++ stream-chat-android-compose/build.gradle
@@ -55,6 +55,7 @@ tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile).all {
 
 dependencies {
     api project(':stream-chat-android-ui-common')
+    implementation project(":stream-chat-android-ui-utils")
 
     implementation Dependencies.androidxAppCompat
