+++ stream-chat-android-compose/detekt-baseline.xml
@@ -33,7 +33,6 @@
     <ID>MagicNumber:ImageAttachmentContent.kt$3</ID>
     <ID>MagicNumber:ImageAttachmentContent.kt$4</ID>
     <ID>MagicNumber:ImagePreviewActivity.kt$ImagePreviewActivity$8f</ID>
-    <ID>MagicNumber:ImageUtils.kt$255</ID>
     <ID>MagicNumber:Messages.kt$3</ID>
     <ID>MagicNumber:Messages.kt$5</ID>
     <ID>MagicNumber:SearchInput.kt$8f</ID>
