+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/util/MessageItemStateUtils.kt
@@ -16,7 +16,6 @@
 
 package io.getstream.chat.android.compose.ui.util
 
-import com.getstream.sdk.chat.utils.extensions.isModerationFailed
 import io.getstream.chat.android.client.utils.SyncStatus
 import io.getstream.chat.android.compose.state.messages.list.MessageItemState
 
@@ -24,8 +23,3 @@ import io.getstream.chat.android.compose.state.messages.list.MessageItemState
  * @return If the current message failed to send.
  */
 internal fun MessageItemState.isFailed(): Boolean = isMine && message.syncStatus == SyncStatus.FAILED_PERMANENTLY
-
-/**
- * @return If the current message failed to send.
- */
-internal fun MessageItemState.isModerationFailed(): Boolean = message.isModerationFailed()
