+++ stream-chat-android-ui-components/build.gradle
@@ -60,6 +60,7 @@ tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile).all {
 
 dependencies {
     api project(":stream-chat-android-ui-common")
+    implementation project(":stream-chat-android-ui-utils")
 
     implementation Dependencies.coroutinesCore
     implementation Dependencies.androidxActivityKtx
