+++ stream-chat-android-ui-components/detekt-baseline.xml
@@ -55,7 +55,6 @@
     <ID>MagicNumber:AvatarView.kt$AvatarView$8f</ID>
     <ID>MagicNumber:ChannelItemSwipeListener.kt$ChannelItemSwipeListener$100</ID>
     <ID>MagicNumber:ChannelViewHolder.kt$ChannelViewHolder$99</ID>
-    <ID>MagicNumber:ColorUtils.kt$255</ID>
     <ID>MagicNumber:EditReactionsView.kt$EditReactionsView$5</ID>
     <ID>MagicNumber:EditReactionsViewStyle.kt$EditReactionsViewStyle.Builder$5</ID>
     <ID>MagicNumber:ImageAttachmentsGroupView.kt$ImageAttachmentsGroupView$3</ID>
