+++ stream-chat-android-ui-utils/detekt-baseline.xml
@@ -1,0 +1,7 @@
+<?xml version='1.0' encoding='UTF-8'?>
+<SmellBaseline>
+  <ManuallySuppressedIssues/>
+  <CurrentIssues>
+    <ID>MagicNumber:ColorUtils.kt$255</ID>
+  </CurrentIssues>
+</SmellBaseline>
