+++ stream-chat-android-ui-utils/src/main/kotlin/io/getstream/chat/android/ui/utils/extensions/String.kt
@@ -14,15 +14,12 @@
  * limitations under the License.
  */
 
-package com.getstream.sdk.chat.utils.extensions
-
-import io.getstream.chat.android.core.internal.InternalStreamChatApi
+package io.getstream.chat.android.ui.utils.extensions
 
 /**
  * Checks if a message contains a link using a
  * regular expression.
  */
-@InternalStreamChatApi
 public fun String.containsLinks(): Boolean {
     val regex =
         """(?:\s|^)((?:https?:)?(?:[a-z|A-Z0-9-]+(?:\.[a-z|A-Z0-9-]+)+)(?::[0-9]+)?(?:(?:[^\s]+)?)?)""".toRegex()
