+++ stream-chat-android-ui-utils/src/test/resources/junit-platform.properties
@@ -1,0 +1,1 @@
+junit.jupiter.testinstance.lifecycle.default = per_class
